feat: shaping, fix: double click

+++ pkg/syntax/form.go
@@ -10,7 +10,7 @@ func TokenizeForm(src []byte) []Token {
 		if start >= end {
 			return
 		}
-		out = append(out, Token{Start: start, End: end, Kind: kind})
+		out = append(out, Token{Start: int32(start), End: int32(end), Kind: kind})
 	}
 
 	i := 0
